# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reverie API",
    "description": "HTTP API for Reverie, a self-hosted ebook library manager. Generated from the server handlers; do not edit by hand.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/books": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List the manifestations visible to the current user, paginated\nvia an opaque cursor and ordered per the requested sort axis.",
        "description": "# Errors\n- [`AppError::MalformedQuery`] when a filter param fails to\n  deserialize (e.g. a malformed UUID in `?author=`/`?series=`/\n  `?shelf=`, or an unknown `?sort=` variant) — HTTP 400 via the\n  `From<QueryRejection>` impl.\n- [`AppError::Validation`] when the cursor is malformed, the sort\n  tag mismatches the cursor, or `tag.len() > MAX_TAG_FILTERS`.\n- [`AppError::Internal`] on database errors.",
        "operationId": "list",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortMode"
            }
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "series",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "shelf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of visible books",
            "headers": {
              "Link": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 8288 next-page link; emitted with rel=\"next\" when more rows remain"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Invalid cursor or too many tag filters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/books/{id}/cover/thumb": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "`GET /api/v1/books/{id}/cover/thumb` — thumbnail cover for the web UI.",
        "description": "# Errors\nSee [`opds_cover`].",
        "operationId": "api_cover_thumb",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Manifestation id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail image stream (`image/jpeg`); Cache-Control: private, max-age=86400; carries a strong ETag"
          },
          "304": {
            "description": "Not Modified — the request's If-None-Match matched the cover ETag"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Manifestation missing, RLS-hidden, or coverless",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "`GET /api/v1/search` — return the top hybrid-ranked manifestations\nmatching `q`, snippet-highlighted via `ts_headline`.",
        "description": "# Errors\n- [`AppError::Validation`] when `q` is missing, empty, or longer\n  than [`MAX_Q_LEN`] characters.\n- [`AppError::Internal`] on database errors.",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Free-form query text. Empty / whitespace-only → 422; over\n[`MAX_Q_LEN`] chars → 422.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top hybrid-ranked search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Empty, missing, or over-length query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/opds/library/search": {
      "get": {
        "tags": [
          "opds"
        ],
        "summary": "`GET /opds/library/search` — full-text search over visible books.",
        "description": "# Errors\n- [`AppError::Internal`] on database errors or unconfigured base URL.",
        "operationId": "library_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search term; an empty (or whitespace-only) query short-circuits\nto an empty feed without hitting the DB.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OPDS acquisition feed of search results; empty/whitespace query yields an empty feed",
            "content": {
              "application/atom+xml;profile=opds-catalog;kind=acquisition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Basic authentication required (WWW-Authenticate: Basic)"
          }
        },
        "security": [
          {
            "opds_basic": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookListResponse": {
        "type": "object",
        "description": "`GET /api/v1/books` response envelope. Carries the page rows plus\nthe opaque cursor for the following page; pagination is also\nsignalled via the RFC 8288 `Link` header on the response.\n\nPrivate to the route module — handler-internal wire shape.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookListRow"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BookListRow": {
        "type": "object",
        "description": "One row of a paginated book list. Decoded via [`sqlx::FromRow`]\nagainst the query in `routes/library::list`, then enriched with\nthe batch-loaded `authors` slot and serialised straight to JSON.",
        "required": [
          "id",
          "work_id",
          "title",
          "authors",
          "cover_url",
          "ingestion_status",
          "validation_status",
          "enrichment_status"
        ],
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Author display names ordered by `work_authors.position`. Empty\nwhen the work has no authors yet (pre-enrichment stub)."
          },
          "cover_url": {
            "type": "string",
            "description": "Cover thumbnail URL — relative path served by the\n`/api/v1/books/{id}/cover/thumb` handler under the caller's\nsession. Not pre-signed; access is gated by the session cookie.\nBackend constructs it so the frontend has a single source of\ntruth for the cover surface."
          },
          "enrichment_status": {
            "$ref": "#/components/schemas/EnrichmentStatus",
            "description": "Enrichment lifecycle state."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "`manifestations.id` — the canonical book id on the wire."
          },
          "ingestion_status": {
            "$ref": "#/components/schemas/IngestionStatus",
            "description": "Ingestion lifecycle state."
          },
          "isbn_13": {
            "type": [
              "string",
              "null"
            ],
            "description": "`manifestations.isbn_13`, when known."
          },
          "series": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SeriesRef",
                "description": "Series membership; `None` when the work isn't on a series."
              }
            ]
          },
          "title": {
            "type": "string",
            "description": "`works.title` of the parent work."
          },
          "validation_status": {
            "$ref": "#/components/schemas/ValidationStatus",
            "description": "Validation lifecycle state."
          },
          "work_id": {
            "type": "string",
            "format": "uuid",
            "description": "`works.id` of the parent work. Lets the client navigate to\n`/api/v1/works/{work_id}` from a list row without an extra fetch."
          }
        }
      },
...
      "EnrichmentStatus": {
        "type": "string",
        "description": "Lifecycle state of metadata enrichment for a single manifestation.\n\nWire-format invariant: variants serialise to the `snake_case` forms\ndeclared in the `#[serde]` and `#[sqlx]` attributes. Drift from\neither side surfaces as a decode error at the boundary, never as a\nsilently coerced unknown value.",
        "enum": [
          "pending",
          "in_progress",
          "complete",
          "failed",
          "skipped"
        ]
      },
...
      "IngestionStatus": {
        "type": "string",
        "description": "Lifecycle state of a single manifestation's ingestion pipeline run.\n\nWire-format invariant: variants serialise to the lowercase forms\ndeclared in the `#[serde]` and `#[sqlx]` attributes; unknown DB\nvariants fail decode loudly instead of coercing into a string.",
        "enum": [
          "pending",
          "processing",
          "complete",
          "failed",
          "skipped"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 `application/problem+json` error body — the single runtime DTO\nfor the error envelope. The `IntoResponse` impl on\n[`crate::error::AppError`] constructs it for every Problem-Details\nvariant, and operational probes (readiness) build it directly, so the\ndocumented component schema and the bytes on the wire come from the same\nstruct and cannot drift.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "URI reference identifying the specific request, when available.\nOmitted (RFC 9457 §3.1 permits this) outside an HTTP request."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "Stable URI reference identifying the problem type.",
            "example": "https://reverie.example/probs/not-found"
          }
        }
      },
...
      "SearchHit": {
        "type": "object",
        "description": "One result row of `GET /api/v1/search`. Carries the bare minimum the\ncommand-palette UI needs: a kind tag for grouping, identifiers for\nnavigation, a short display label, and an optional `ts_headline`\nsnippet with non-HTML ASCII STX (`\\u{0002}`) / ETX (`\\u{0003}`)\nmarkers so the React renderer can avoid `dangerouslySetInnerHTML`.",
        "required": [
          "kind",
          "id",
          "title",
          "authors"
        ],
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Author display names for `book` results."
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cover thumbnail URL for `book` results — relative path under\n`/api/v1/books/{id}/cover/thumb`, session-cookie gated. `None`\nwhen the hit kind has no cover surface (future author/series\nvariants)."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Primary id — `manifestations.id` for `book`, `authors.id` for\n`author`, `series.id` for `series`."
          },
          "kind": {
            "$ref": "#/components/schemas/SearchHitKind",
            "description": "Result kind. Currently always `\"book\"`; `\"author\"` and\n`\"series\"` ship in a follow-up that fans the hybrid CTE over\nthe existing `authors.name` / `series.name` trigram indexes."
          },
          "snippet": {
            "type": [
              "string",
              "null"
            ],
            "description": "`ts_headline` snippet from the work's title+description with\nASCII STX (`\\u{0002}`) / ETX (`\\u{0003}`) start/stop markers\naround matched terms. Control codepoints, not valid UTF-8 text,\nso they cannot collide with user typography. `None` when the\nhit was trigram-only (no tsquery match → headline would be the\nraw text without highlighting)."
          },
          "title": {
            "type": "string",
            "description": "Display label — work title for `book`."
          },
          "work_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Parent work id when `kind = \"book\"`, else `None`."
          }
        }
      },
      "SearchHitKind": {
        "type": "string",
        "description": "Tag identifying which entity a [`SearchHit`] points at. Serialised\nin `snake_case` to match the rest of the JSON API conventions.",
        "enum": [
          "book"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "description": "`GET /api/v1/search` envelope (11b). Wraps a flat result list — the\nfrontend groups by [`SearchHit::kind`] client-side. No cursor:\nsearch is bounded by `LIMIT` server-side; pagination is a follow-up\nif user-research warrants it.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "description": "Result rows, ranked DESC by hybrid `ts_rank_cd + similarity`."
          }
        }
      },
...
      "SeriesRef": {
        "type": "object",
        "description": "Series membership for a manifestation. Embedded into both\n[`BookListRow`] and [`BookDetail`]; `None` when the work isn't on\nany series.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Series primary key."
          },
          "name": {
            "type": "string",
            "description": "Series display name."
          },
          "position": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Position within the series (`series_works.position`), `None`\nwhen the membership row has a null position."
          }
        }
      },
...
      "SortMode": {
        "type": "string",
        "description": "Sort axis selectable via `?sort=...` on `/api/v1/books`.",
        "enum": [
          "recent",
          "title",
          "author"
        ]
      },
...
      "ValidationStatus": {
        "type": "string",
        "description": "Outcome of structural validation for a single manifestation.\n\nWire-format invariant: variants serialise to the lowercase forms declared\nin the `#[serde]` and `#[sqlx]` attributes. `clean`, `repaired`, and\n`degraded` are all stored-and-served outcomes on one quality tier — `clean`\nmeans *no issues found*, not *the only valid state*. Unknown DB variants\nfail decode loudly at the boundary instead of coercing into a string.\n\n`#[non_exhaustive]` is deliberately absent: `reverie_api` is a single\nlib+bin crate with no foreign downstream matching on this enum, and the\nloud-decode-failure design *wants* the compile error when a\n[`ValidationOutcome`](crate::services::epub::ValidationOutcome) variant is\nadded without a matching arm — `non_exhaustive` would suppress exactly that.",
        "enum": [
          "pending",
          "clean",
          "repaired",
          "degraded",
          "failed"
        ]
      },
...
    },
    "securitySchemes": {
      "opds_basic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic authentication for the OPDS feeds. MUST be used over HTTPS in production — Basic credentials are otherwise exposed in transit (Checkov CKV_OPENAPI_3)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id",
        "description": "Session cookie issued by the server's session layer. Set with the Secure attribute when `behind_https` is enabled; MUST be served over HTTPS in production to prevent session hijacking."
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    }
  ],
...
}